// AuctionPLL/src/app/generated-models/comment-view-model.ts
/* eslint-disable */
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export class CommentViewModel {
    id: string;
    author: string;
    text: string;
    dateTime: Date;
    lotId: string;
    userId: string;
    isBid: boolean;
}